chore: reorganise components
