+++ src/components/Buttons/index.js
@@ -1,4 +1,2 @@
 export { default as DialogButton } from './DialogButton';
-export { default as ActionButtons } from './ActionButtons';
-export { default as AddButton } from './AddButton';
 export { default as HelpButton } from './HelpButton';
