+++ src/components/JobFields/index.js
@@ -1,5 +1,6 @@
 import Type from './Type';
 import Name from './Name';
 import Cron from './Cron';
+import ActionButtons from './ActionButtons';
 
-export { Type, Name, Cron };
+export { Type, Name, Cron, ActionButtons };
